+++ src/libtrx/include/libtrx/game/gun/control.h
@@ -8,3 +8,5 @@ void Gun_UpdateGunState(void);
 int32_t Gun_FireWeapon(
     LARA_GUN_TYPE weapon_type, ITEM *target, const ITEM *src,
     const int16_t *angles);
+
+void Gun_Control(void);