+++ src/libtrx/include/libtrx/game/gun/pistols.h
@@ -4,6 +4,8 @@
 
 extern void Gun_Pistols_SetArmInfo(LARA_ARM *arm, int32_t frame);
 
+extern void Gun_Pistols_Control(LARA_GUN_TYPE weapon_type);
+
 void Gun_Pistols_Draw(LARA_GUN_TYPE weapon_type);
 void Gun_Pistols_Undraw(LARA_GUN_TYPE weapon_type);
 void Gun_Pistols_Ready(LARA_GUN_TYPE weapon_type);