+++ src/tr1/game/gun/gun_pistols.h
@@ -4,8 +4,4 @@
 //
 // This refers to the Pistols, but also the Magnums and Uzis.
 
-#include "global/types.h"
-
 #include <libtrx/game/gun/pistols.h>
-
-void Gun_Pistols_Control(LARA_GUN_TYPE weapon_type);