+++ src/tr1/game/gun/gun_rifle.h
@@ -5,12 +5,7 @@
 // In Tomb Raider 1 this means only shotgun. Future games add more weapons such
 // as M-16, Harpoon, Grenade Launcher etc.
 
-#include "global/types.h"
-
 #include <libtrx/game/gun/rifle.h>
 
-void Gun_Rifle_Draw(LARA_GUN_TYPE weapon_type);
-void Gun_Rifle_Undraw(LARA_GUN_TYPE weapon_type);
-void Gun_Rifle_Ready(LARA_GUN_TYPE weapon_type);
-void Gun_Rifle_Control(LARA_GUN_TYPE weapon_type);
 void Gun_Rifle_Animate(LARA_GUN_TYPE weapon_type);
+void Gun_Rifle_Ready(LARA_GUN_TYPE weapon_type);