+++ src/tr2/game/lara/flare.h
@@ -1,8 +1,3 @@
 #pragma once
 
 #include <libtrx/game/lara/flare.h>
-
-void Lara_Flare_Control(void);
-void Lara_Flare_Draw(void);
-void Lara_Flare_Undraw(void);
-void Lara_Flare_Dispose(bool thrown);